All ThingExtensions method moved: Fix #364
