+++ COMET.Web.Common/COMET.Web.Common.csproj
@@ -2,7 +2,7 @@
     <PropertyGroup>
         <TargetFramework>net7.0</TargetFramework>
         <LangVersion>Latest</LangVersion>
-        <Version>1.0.6</Version>
+        <Version>1.0.7</Version>
         <AssemblyVersion>1.0.0</AssemblyVersion>
         <FileVersion>1.0.0</FileVersion>
         <Title>CDP4 WEB Common</Title>
