+++ COMETwebapp/COMETwebapp.csproj
@@ -74,7 +74,7 @@
     <ItemGroup>
         <PackageReference Include="AntDesign" Version="0.14.1" />
         <PackageReference Include="BlazorStrap" Version="5.1.100" />
-        <PackageReference Include="CDP4.WEB.Common" Version="1.0.5" />
+        <PackageReference Include="CDP4.WEB.Common" Version="1.0.7" />
         <PackageReference Include="System.Drawing.Common" Version="7.0.0" />
         <PackageReference Include="System.Drawing.Primitives" Version="4.3.0" />
         <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="7.0.4" />
